refactor(network): export everything on index
